+++ gossip/src/node.rs
@@ -11,13 +11,11 @@ use {
         multihomed_sockets::BindIpAddrs,
         sockets::{
             bind_gossip_port_in_range, bind_in_range_with_config, bind_more_with_config,
-            bind_to_with_config, bind_two_in_range_with_offset_and_config,
-            localhost_port_range_for_tests, multi_bind_in_range_with_config,
+            bind_to_with_config, localhost_port_range_for_tests, multi_bind_in_range_with_config,
             SocketConfiguration as SocketConfig,
         },
     },
     solana_pubkey::Pubkey,
-    solana_quic_definitions::QUIC_PORT_OFFSET,
     solana_streamer::quic::DEFAULT_QUIC_ENDPOINTS,
     solana_time_utils::timestamp,
     std::{
@@ -134,18 +132,15 @@ impl Node {
             bind_in_range_with_config(bind_ip_addr, port_range, socket_config)
                 .expect("tvu_quic bind");
 
-        let ((tpu_port, tpu_socket), (tpu_port_quic, tpu_quic)) =
-            bind_two_in_range_with_offset_and_config(
-                bind_ip_addr,
-                port_range,
-                QUIC_PORT_OFFSET,
-                socket_config,
-                socket_config,
-            )
-            .expect("tpu_socket primary bind");
+        let (tpu_port, tpu_socket) =
+            bind_in_range_with_config(bind_ip_addr, port_range, socket_config)
+                .expect("tpu_socket primary bind");
         let tpu_sockets =
             bind_more_with_config(tpu_socket, 32, socket_config).expect("tpu_sockets multi_bind");
 
+        let (tpu_port_quic, tpu_quic) =
+            bind_in_range_with_config(bind_ip_addr, port_range, socket_config)
+                .expect("tpu_socket primary bind");
         let mut tpu_quic = bind_more_with_config(tpu_quic, num_quic_endpoints.get(), socket_config)
             .expect("tpu_quic bind");
 
@@ -156,17 +151,14 @@ impl Node {
         );
         let tpu_quic_addresses = Self::get_socket_addrs(&tpu_quic);
 
-        let ((tpu_forwards_port, tpu_forwards_socket), (tpu_forwards_quic_port, tpu_forwards_quic)) =
-            bind_two_in_range_with_offset_and_config(
-                bind_ip_addr,
-                port_range,
-                QUIC_PORT_OFFSET,
-                socket_config,
-                socket_config,
-            )
-            .expect("tpu_forwards primary bind");
+        let (tpu_forwards_port, tpu_forwards_socket) =
+            bind_in_range_with_config(bind_ip_addr, port_range, socket_config)
+                .expect("tpu_forwards primary bind");
         let tpu_forwards_sockets = bind_more_with_config(tpu_forwards_socket, 8, socket_config)
             .expect("tpu_forwards multi_bind");
+        let (tpu_forwards_quic_port, tpu_forwards_quic) =
+            bind_in_range_with_config(bind_ip_addr, port_range, socket_config)
+                .expect("tpu_forwards primary bind");
         let mut tpu_forwards_quic =
             bind_more_with_config(tpu_forwards_quic, num_quic_endpoints.get(), socket_config)
                 .expect("tpu_forwards_quic multi_bind");
@@ -303,9 +295,16 @@ impl Node {
         )
         .unwrap();
         info.set_tvu(QUIC, (advertised_ip, tvu_quic_port)).unwrap();
-        info.set_tpu(public_tpu_addr.unwrap_or_else(|| SocketAddr::new(advertised_ip, tpu_port)))
+        info.set_tpu(UDP, (advertised_ip, tpu_port)).unwrap();
+        info.set_tpu(
+            QUIC,
+            public_tpu_addr.unwrap_or_else(|| SocketAddr::new(advertised_ip, tpu_port_quic)),
+        )
+        .unwrap();
+        info.set_tpu_forwards(UDP, (advertised_ip, tpu_forwards_port))
             .unwrap();
         info.set_tpu_forwards(
+            QUIC,
             public_tpu_forwards_addr
                 .unwrap_or_else(|| SocketAddr::new(advertised_ip, tpu_forwards_port)),
         )
@@ -485,13 +484,13 @@ mod multihoming {
             // tpu_quic
             let tpu_quic_address = self.addresses.tpu_quic[interface_index];
             cluster_info
-                .set_tpu(tpu_quic_address)
+                .set_tpu_quic(tpu_quic_address)
                 .map_err(|e| e.to_string())?;
 
             // tpu_forwards_quic
             let tpu_forwards_quic_address = self.addresses.tpu_forwards_quic[interface_index];
             cluster_info
-                .set_tpu_forwards(tpu_forwards_quic_address)
+                .set_tpu_forwards_quic(tpu_forwards_quic_address)
                 .map_err(|e| e.to_string())?;
 
             // tpu_vote_quic