+++ vortexor/src/main.rs
@@ -6,7 +6,6 @@ use {
     solana_core::banking_trace::BankingTracer,
     solana_keypair::read_keypair_file,
     solana_net_utils::sockets::{bind_in_range_with_config, SocketConfiguration as SocketConfig},
-    solana_quic_definitions::QUIC_PORT_OFFSET,
     solana_signer::Signer,
     solana_streamer::streamer::StakedNodes,
     solana_vortexor::{
@@ -22,7 +21,7 @@ use {
     std::{
         collections::HashMap,
         env,
-        net::{IpAddr, SocketAddr},
+        net::IpAddr,
         path::PathBuf,
         sync::{atomic::AtomicBool, Arc, RwLock},
     },
@@ -169,16 +168,8 @@ pub fn main() {
         tpu_sockets.tpu_quic_fwd[0].local_addr()
     );
 
-    let tpu_address = tpu_sockets.tpu_quic[0].local_addr().unwrap();
-    let tpu_public_address = SocketAddr::new(
-        tpu_address.ip(),
-        tpu_address.port().saturating_sub(QUIC_PORT_OFFSET),
-    );
-    let tpu_fwd_address = tpu_sockets.tpu_quic_fwd[0].local_addr().unwrap();
-    let tpu_fwd_public_address = SocketAddr::new(
-        tpu_fwd_address.ip(),
-        tpu_fwd_address.port().saturating_sub(QUIC_PORT_OFFSET),
-    );
+    let tpu_public_address = tpu_sockets.tpu_quic[0].local_addr().unwrap();
+    let tpu_fwd_public_address = tpu_sockets.tpu_quic_fwd[0].local_addr().unwrap();
 
     for destination in destinations.read().unwrap().iter() {
         info!(